Loki: Split scheduler tests into themed modules
Break the monolithic scheduler test file into graph, dependency, processing and config-focused modules. Move shared fixtures and visualisation helpers into dedicated test support files so the coverage is easier to navigate and maintain.
